requirements_fallback.txt
streamlit>=1.28.1,<2.0.0
numpy>=1.24.3,<2.0.0
pandas>=1.5.3,<2.0.0
pillow>=10.0.1,<11.0.0
matplotlib>=3.7.2,<4.0.0
scikit-learn>=1.3.0,<2.0.0
nltk>=3.8.1,<4.0.0
requests>=2.31.0,<3.0.0
